# openapi.yml
openapi: '3.0.2'
info:
  title: Clique Ledger RESTAPI
  version: '1.0'
servers:
  - url: https://api.cliquesledger.com/v1
...
paths:
  /cliques:
    get:
      tags:
        - Clique 
      responses:
        200:
          description: Cliques
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/CliqueResponse'

    post:
      tags:
        - Clique 
      description: This is to create a cliques
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CliqueRequest'
              
      responses:
        201:
          description: Clique created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CliqueResponse'
  /cliques/{cliqueId}:
    get:
      tags: 
        - Clique
      parameters:
        - $ref: '#/components/parameters/CliqueIdParam' 
      responses:
        200:
          description: To get a cliques of given cliques id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliqueResponse"
    patch:
      tags: 
        - Clique
      parameters:
        - $ref: '#/components/parameters/CliqueIdParam'
      requestBody:
        description: 
          This contains some partial update of cliquess. Eg. Clique name
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                  example: cliques-1
      responses:
        200:
          description: This is when the property is updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  field_name: 
                    type: string
                    example: name
                  status:
                    type: string
                    example: SUCCESS
        400:
          description: This is when someone sends invalid field
          content:
            application/json:
              schema:
                type: object
                properties:
                  field_name: 
                    type: string
                    example: name
                  status:
                    type: string
                    example: FAILED
        404:
          description: This is when someone sends invalid cliquesId
          content:
            application/json:
              schema:
                type: object
                properties:
                  field_name: 
                    type: string
                    example: name
                  status:
                    type: string
                    example: NOT FOUND
    delete:
      tags:
        - Clique
      parameters:
        - $ref: '#/components/parameters/CliqueIdParam'

      description: This is to delete a cliques 
      responses:
        204: 
          description: Successfully Deleted the cliques 
        404: 
          description: In case Invalid id is provided for cliques
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: NOT FOUND
  /cliques/{cliquesId}/members/:
    post:
      description: To add members in the group
      tags: 
          - Clique
      parameters:
        - $ref: '#/components/parameters/CliqueIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              example:
                - U00001
                - U00002
                - U00003
      responses:
        201:
          description: Members added successfully. This response will return the member id for each user id provided in the request array.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  status: 
                    type: string
                    example: SUCCESS
                  message: 
                    type: string
                    example: members added successfully
                  member_ids: 
                    example:
                      - uid: U00001
                        mid: M000001
                      - uid: U00002
                        mid: M000002
                      - uid: U00003
                        mid: M000003
        400:
          description: In case one or more invalid userid is provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: FAILED
                  message:
                    type: string
                    example: One or more userid is invalid. No members added. 
    delete:
      description: To remove members from the group
      tags: 
          - Clique
      parameters:
        - $ref: '#/components/parameters/CliqueIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              example: 
                 - M00002
                 - M00003
      responses:
        200:
          description: members removed successfully
          content:
            application/json:
              schema: 
                type: object
                properties:
                  status: 
                    type: string
                    example: SUCCESS
                  message: 
                    type: string
                    example: members added successfully
        400:
          description: In case one or more invalid member is provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: 
                    type: string
                    example: FAILED
                  message:
                    type: string
                    example: One or more userid is invalid.  No mebers deleted. 
...
components:
...
  parameters:
    CliqueIdParam:
        in: path
        name: cliqueId
        description: This contains cliques Id
        required: true
        schema:
          type: string
          example: L000001
...
  schemas:
    CliqueRequest:
      type: object
      properties:
        name: 
          type: string
          example: Alexa
        funds:
          type: number
          example: 5000
    CliqueResponse:
      type: object
      properties:
        clique_id:
          type: string
          example: C000001
        name:
          type: string
          example: clique name
        members:
          type: array
          items:
            type: object
            properties:
              user_id:
                type: string
                example: U00001
              member_id:
                type: string
                example: M00001
              member_name:
                type: string
                example: Member Name
              is_admin:
                type: boolean
                example: true
              
        is_fund:
          type: boolean
          example: true
        fund:
          type: number
          example: 5000
        isActive:
          type: boolean
          example: true
...